requirements.txt
absl-py==1.4.0
aiofiles 
aiohappyeyeballs 
aiohttp
aiosignal
alphafold-colabfold==2.3.8
altair
annotated-types
anyio
appdirs
astor
asttokens==2.2.1
astunparse
async-timeout
attrs
backcall==0.2.0
base58
biopython
blinker
boto3
botocore
Brotli
cached-property
cachetools==5.5.2
certifi
cffi
charset-normalizer
chex==0.1.89
click
colabfold @ git+https://github.com/sokrypton/ColabFold@8c55fc2548823d57bb23c24a6b2573348e6d51c8
colorama
comm==0.1.3
contourpy
cycler
debugpy==1.6.7
decorator==5.1.1
dm-haiku==0.0.13
dm-tree==0.1.9
dnspython
email_validator
etils==1.12.2
eval_type_backport==0.2.2
exceptiongroup
executing==1.2.0
fastapi
fastapi-cli
filelock
flatbuffers
flax==0.10.6
fonttools
frozenlist
fsspec==2025.5.1
gast
gitdb
GitPython
google-pasta
gradio_client==1.10.3
greenlet
grpcio
h11
h2
h5py
hpack
httpcore
httptools
httpx
huggingface-hub==0.33.0
humanize==4.12.3
hyperframe
idna
immutabledict==4.2.1
importlib_metadata
importlib_resources
ipykernel==6.23.1
ipython==8.13.2
jax==0.6.1
jaxlib==0.6.1
jedi==0.18.2
Jinja2
jmespath
jmp==0.0.4
joblib
jsonschema
jsonschema-specifications
jupyter_client==8.2.0
jupyter_core==5.3.0
keras
kiwisolver
Markdown
markdown-it-py
MarkupSafe
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdurl
mistralai==1.8.2
ml-dtypes
ml_collections==1.1.0
msgpack==1.1.0
multidict
munkres==1.1.4
namex
narwhals
nest-asyncio==1.6.0
numpy==2.0.2 
OpenMM==8.2.0
opt_einsum
optax==0.2.5
optree
orbax-checkpoint==0.11.14
packaging==24.2
pandas==2.2.2
parso==0.8.3
pickleshare==0.7.5
Pillow 
pkgutil_resolve_name
prompt-toolkit==3.0.38
propcache
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
py3Dmol
pyarrow
pycparser
pydantic
pydantic_core
pydeck
Pygments
Pympler
pyparsing
pyperclip==1.9.0
PySide6==6.7.2
PySocks
python-dateutil
python-dotenv
python-multipart
pytz
pywin32==306
PyYAML
pyzmq==25.0.2
referencing
requests
rich
rich-toolkit
rpds-py
s3transfer
scikit-learn 
scipy 
semver
shellingham
shiboken6==6.7.2
simplejson==3.20.1
six
smmap
sniffio
SQLAlchemy
stack-data==0.6.2
starlette
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard
tensorboard_data_server
tensorflow_cpu==2.18.1
tensorstore==0.1.75
termcolor
threadpoolctl
toml
toolz==1.0.0
tornado==6.3.2
tqdm
traitlets==5.9.0
treescope==0.1.9
typer==0.16.0
typer-slim==0.16.0
typing-inspection
typing_extensions
tzdata
tzlocal
unicodedata2
urllib3
uvicorn
validators
watchdog
watchfiles
wcwidth==0.2.6
websockets
Werkzeug
win_inet_pton
wrapt
yarl
zipp
zstandard==0.23.0